+++ Common/pom.xml
@@ -11,11 +11,6 @@
     <artifactId>Common</artifactId>
 
     <dependencies>
-        <dependency>
-            <groupId>jakarta.persistence</groupId>
-            <artifactId>jakarta.persistence-api</artifactId>
-            <version>3.1.0</version>
-        </dependency>
     </dependencies>
 
     <build>
