--- Common/src/main/java/edu/sdccd/cisc191/common/entities/TestAdminUser.java
@@ -1,15 +1,0 @@
-package edu.sdccd.cisc191.common.entities;
-
-public class TestAdminUser {
-    public static void main(String[] args){
-        Stock apple = new Stock("AAPL");
-        Stock amazon = new Stock("AMZN");
-        Stock microsoft = new Stock("MSFT");
-        Stock[] myStockList = {apple,amazon,microsoft};
-        StockList techStocks = new StockList(myStockList);
-        String totalOutput = techStocks.getAllStockInfo();
-        System.out.println(totalOutput);
-    }
-
-    public TestAdminUser(){}
-}
