+++ Server/src/main/java/edu/sdccd/cisc191/server/repositories/StockRepository.java
@@ -1,7 +1,10 @@
 package edu.sdccd.cisc191.server.repositories;
 
-import edu.sdccd.cisc191.common.entities.Stock;
 import org.springframework.data.repository.CrudRepository;
 
 public interface StockRepository extends CrudRepository<Stock, String> {
+
+    public static void main(String[] args){
+        Stock myStock = new Stock();
+    }
 }
